Split js file into multiple modules
