+++ index.html
@@ -34,6 +34,6 @@
         </div>
         
         <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-        <script src="script.js" async defer></script>
+        <script type="module" src="script.js" async defer></script>
     </body>
 </html>
